Added writer and loader for condensed OBJ file format
Added writer and loader for condensed OBJ file format

Loading waverfront OBJ-files takes too long as it is a plain-text format.
I wrote a program that stores the loaded data into a binary representation,
using this utility class. The speedup is anywhere between 50x to 100x.
The Rungholt scene load time went down from 36 to 0.6 seconds.

Only a subset of the OBJ format is supported, and better error detection needs
to be written. For now, it is nice that loading a complex scene takes a fraction
of the time.
    src/util/CObjUtil.cpp (+)
    src/util/CObjUtil.h   (+)